Bugfix: organized components into folders
