+++ src/App.jsx
@@ -1,6 +1,7 @@
 // import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Navbar from './component/Navbar';
+import Navbar from './component/navbar/Navbar';
+
 import Home from './pages/Home';
 import About from './pages/AboutTheSite';
 import VolunteerSearch from './pages/VolunteerSearch';
@@ -10,7 +11,7 @@ import UserPage from './pages/UserPage';
 import EcoChat from './pages/Eco-Chat';
 import AboutTheCreators from './pages/AboutTheCreators';
 import SiteSponsors from './pages/SiteSponsors.jsx';
-import Footer from './component/Footer.jsx';
+import Footer from './component/footer/Footer';
 
 function App() {
   return (
