+++ src/component/search-bar/search_bar.module.scss
